requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-tinymce==4.0.0
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
style==1.1.0
update==0.0.1
urllib3==2.2.1
